Crystal currently loops forever running an analysis on the following code:
private LabeledResult(List<ILabel> labels,LE defaultValue){
 labelMap=new HashMap<ILabel,LE>();
 this.defaultValue=defaultValue;
 for (  ILabel lab : labels) {
   this.labelMap.put(lab,defaultValue);       ***** dying on the this in this statement
 }
}


Specifically, Crystal tries to create a ThisVariable inside the loop.
It first looks to see if a variable has previously been created, finds
none, and creates a new one. It does this on every iteration of the
loop, meaning the results keep growing with a new ThisVariable each
time through. The problem seems to be in EclipseTAC.getThisVariable():


       private ThisVariable getThisVariable(ThisExpression node) {
               Name qualifier = node.getQualifier();
               /**
                * The rule is that the declared, generic type, is the single canonical type.
                * Parameterized types are actually different type bindings.
                * @see org.eclipse.jdt.internal.compiler.lookup.LookupEnvironment#createParameterizedType(ReferenceBinding
genericType, TypeBinding[] typeArguments, ReferenceBinding
enclosingType)
                */
               ITypeBinding generic_type =
node.resolveTypeBinding().getTypeDeclaration();    **** get type
declaration of binding
               ThisVariable result = thisVar.get(generic_type);
               if(result == null) {
                       // explicitly qualified this
                       result = new ThisVariable(this, qualifier);
                       thisVar.put(node.resolveTypeBinding(), result);              ******
BUG - but write out the binding
               }
               if(result.isImplicit()) {
                       // fix up qualifier of previously created implicitly qualified this variable
                       result.explicitQualifier(node.getQualifier());
               }
               return result;
       }


This code is caching ThisVariables and lazily creating them. But, it
looks them up based on the type declaration of the binding but stores
them using the binding. When they're the same, this works. Otherwise,
it doesn't and we create a new variable. Changing it to be consistent

                       thisVar.put(node.resolveTypeBinding().getTypeDeclaration(), result);

fixes this bug.